+++ lib/src/core/toastification_manager.dart
@@ -39,6 +39,7 @@ class ToastificationManager {
     required Duration? animationDuration,
     Duration? autoCloseDuration,
     OverlayState? overlayState,
+    VoidCallback? onAutoCompleteCompleted,
   }) {
     final item = ToastificationItem(
       builder: builder,
@@ -48,6 +49,7 @@ class ToastificationManager {
       autoCloseDuration: autoCloseDuration,
       onAutoCompleteCompleted: (holder) {
         dismiss(holder);
+        onAutoCompleteCompleted?.call();
       },
     );